+++ aea/registries/base.py
@@ -22,32 +22,22 @@
 import importlib.util
 import inspect
 import logging
-import os
 import pprint
-import queue
 import re
 from abc import ABC, abstractmethod
 from pathlib import Path
-from queue import Queue
-from typing import Dict, Generic, List, Optional, Set, Tuple, TypeVar, Union, cast
+from typing import Dict, Generic, List, Optional, Set, Tuple, TypeVar, cast
 
 from aea.configurations.base import (
-    AgentConfig,
-    ConfigurationType,
-    DEFAULT_AEA_CONFIG_FILE,
     DEFAULT_PROTOCOL_CONFIG_FILE,
-    DEFAULT_SKILL_CONFIG_FILE,
     ProtocolConfig,
     ProtocolId,
     PublicId,
     SkillId,
 )
 from aea.configurations.loader import ConfigLoader
-from aea.decision_maker.messages.base import InternalMessage
-from aea.decision_maker.messages.transaction import TransactionMessage
-from aea.mail.base import EnvelopeContext
-from aea.protocols.base import Message, Protocol
-from aea.skills.base import AgentContext, Behaviour, Handler, Model, Skill
+from aea.protocols.base import Protocol
+from aea.skills.base import Behaviour, Handler, Model
 from aea.skills.tasks import Task
 
 logger = logging.getLogger(__name__)
@@ -531,297 +521,3 @@ class HandlerRegistry(ComponentRegistry[Handler]):
         :return: the internal handler registered for the skill id
         """
         return self.fetch_by_protocol_and_skill(INTERNAL_PROTOCOL_ID, skill_id)
-
-
-class Resources:
-    """This class implements the resources of an AEA."""
-
-    def __init__(self, directory: Optional[Union[str, os.PathLike]] = None):
-        """Instantiate the resources."""
-        self._directory = (
-            str(Path(directory).absolute())
-            if directory is not None
-            else str(Path(".").absolute())
-        )
-        self.protocol_registry = ProtocolRegistry()
-        self.handler_registry = HandlerRegistry()
-        self.behaviour_registry = ComponentRegistry[Behaviour]()
-        self.model_registry = ComponentRegistry[Model]()
-        self._skills = dict()  # type: Dict[SkillId, Skill]
-
-        self._registries = [
-            self.protocol_registry,
-            self.handler_registry,
-            self.behaviour_registry,
-            self.model_registry,
-        ]
-
-    @property
-    def directory(self) -> str:
-        """Get the directory."""
-        return self._directory
-
-    def _load_agent_config(self) -> AgentConfig:
-        """Load the agent configuration."""
-        config_loader = ConfigLoader.from_configuration_type(ConfigurationType.AGENT)
-        agent_config = config_loader.load(
-            open(os.path.join(self.directory, DEFAULT_AEA_CONFIG_FILE))
-        )
-        return agent_config
-
-    def load(self, agent_context: AgentContext) -> None:
-        """Load all the resources."""
-        agent_configuration = self._load_agent_config()
-        self.protocol_registry.populate(
-            self.directory, allowed_protocols=agent_configuration.protocols
-        )
-        self.populate_skills(
-            self.directory, agent_context, allowed_skills=agent_configuration.skills
-        )
-
-    def populate_skills(
-        self,
-        directory: str,
-        agent_context: AgentContext,
-        allowed_skills: Optional[Set[PublicId]] = None,
-    ) -> None:
-        """
-        Populate skills.
-
-        :param directory: the agent's resources directory.
-        :param agent_context: the agent's context object
-        :param allowed_skills: an optional set of allowed skills (public ids).
-                               If None, every skill is allowed.
-        :return: None
-        """
-        skill_directory_paths = set()  # type: ignore
-
-        # find all skill directories from vendor/*/skills
-        skill_directory_paths.update(Path(directory, "vendor").glob("./*/skills/*/"))
-        # find all skill directories from skills/
-        skill_directory_paths.update(Path(directory, "skills").glob("./*/"))
-
-        skills_packages_paths = list(
-            filter(
-                lambda x: PACKAGE_NAME_REGEX.match(str(x.name)) and x.is_dir(),
-                skill_directory_paths,
-            )
-        )  # type: ignore
-        logger.debug(
-            "Found the following skill packages: {}".format(
-                pprint.pformat(map(str, skills_packages_paths))
-            )
-        )
-        for skill_directory in skills_packages_paths:
-            logger.debug(
-                "Processing the following skill directory: '{}".format(skill_directory)
-            )
-            try:
-                skill_loader = ConfigLoader.from_configuration_type(
-                    ConfigurationType.SKILL
-                )
-                skill_config = skill_loader.load(
-                    open(skill_directory / DEFAULT_SKILL_CONFIG_FILE)
-                )
-                if (
-                    allowed_skills is not None
-                    and skill_config.public_id not in allowed_skills
-                ):
-                    logger.debug(
-                        "Ignoring skill {}, not declared in the configuration file.".format(
-                            skill_config.public_id
-                        )
-                    )
-                    continue
-                else:
-                    skill = Skill.from_dir(str(skill_directory), agent_context)
-                    assert skill is not None
-                    self.add_skill(skill)
-            except Exception as e:
-                logger.warning(
-                    "A problem occurred while parsing the skill directory {}. Exception: {}".format(
-                        skill_directory, str(e)
-                    )
-                )
-
-    def add_skill(self, skill: Skill):
-        """Add a skill to the set of resources."""
-        skill_id = skill.config.public_id
-        self._skills[skill_id] = skill
-        if skill.handlers is not None:
-            for handler in skill.handlers.values():
-                self.handler_registry.register((skill_id, handler.name), handler)
-        if skill.behaviours is not None:
-            for behaviour in skill.behaviours.values():
-                self.behaviour_registry.register((skill_id, behaviour.name), behaviour)
-        if skill.models is not None:
-            for model in skill.models.values():
-                self.model_registry.register((skill_id, model.name), model)
-
-    def add_protocol(self, protocol: Protocol):
-        """Add a protocol to the set of resources."""
-        self.protocol_registry.register(protocol.id, protocol)
-
-    def get_skill(self, skill_id: SkillId) -> Optional[Skill]:
-        """Get the skill."""
-        return self._skills.get(skill_id, None)
-
-    def get_all_skills(self) -> List[Skill]:
-        """
-        Get the list of all the skills.
-
-        :return: the list of skills.
-        """
-        return list(self._skills.values())
-
-    def remove_skill(self, skill_id: SkillId):
-        """Remove a skill from the set of resources."""
-        self._skills.pop(skill_id, None)
-        try:
-            self.handler_registry.unregister_by_skill(skill_id)
-        except ValueError:
-            pass
-
-        try:
-            self.behaviour_registry.unregister_by_skill(skill_id)
-        except ValueError:
-            pass
-
-    def setup(self):
-        """
-        Set up the resources.
-
-        :return: None
-        """
-        for r in self._registries:
-            r.setup()
-
-    def teardown(self):
-        """
-        Teardown the resources.
-
-        :return: None
-        """
-        for r in self._registries:
-            r.teardown()
-
-
-class Filter:
-    """This class implements the filter of an AEA."""
-
-    def __init__(self, resources: Resources, decision_maker_out_queue: Queue):
-        """
-        Instantiate the filter.
-
-        :param resources: the resources
-        :param decision_maker_out_queue: the decision maker queue
-        """
-        self._resources = resources
-        self._decision_maker_out_queue = decision_maker_out_queue
-
-    @property
-    def resources(self) -> Resources:
-        """Get resources."""
-        return self._resources
-
-    @property
-    def decision_maker_out_queue(self) -> Queue:
-        """Get decision maker (out) queue."""
-        return self._decision_maker_out_queue
-
-    def get_active_handlers(
-        self, protocol_id: PublicId, envelope_context: Optional[EnvelopeContext]
-    ) -> List[Handler]:
-        """
-        Get active handlers.
-
-        :param protocol_id: the protocol id
-        :param envelope context: the envelope context
-        :return: the list of handlers currently active
-        """
-        skill_id = None  # Optional[PublicId]
-        if envelope_context is not None and envelope_context.uri is not None:
-            uri_path = envelope_context.uri.path
-            try:
-                skill_id = PublicId.from_uri_path(uri_path)
-            except ValueError:
-                logger.warning("URI - {} - not a valid skill id.".format(uri_path))
-
-        if skill_id is not None:
-            handler = self.resources.handler_registry.fetch_by_protocol_and_skill(
-                protocol_id, skill_id
-            )
-            active_handlers = (
-                [] if handler is None or not handler.context.is_active else [handler]
-            )
-        else:
-            handlers = self.resources.handler_registry.fetch_by_protocol(protocol_id)
-            active_handlers = list(
-                filter(lambda handler: handler.context.is_active, handlers)
-            )
-        return active_handlers
-
-    def get_active_behaviours(self) -> List[Behaviour]:
-        """
-        Get the active behaviours.
-
-        :return: the list of behaviours currently active
-        """
-        behaviours = self.resources.behaviour_registry.fetch_all()
-        active_behaviour = list(
-            filter(lambda b: b.context.is_active and not b.is_done(), behaviours,)
-        )
-        return active_behaviour
-
-    def handle_internal_messages(self) -> None:
-        """
-        Handle the messages from the decision maker.
-
-        :return: None
-        """
-        while not self.decision_maker_out_queue.empty():
-            try:
-                internal_message = (
-                    self.decision_maker_out_queue.get_nowait()
-                )  # type: Optional[InternalMessage]
-            except queue.Empty:
-                logger.warning("The decision maker out queue is unexpectedly empty.")
-                continue
-            if internal_message is None:
-                logger.warning("Got 'None' while processing internal messages.")
-            elif isinstance(internal_message, TransactionMessage):
-                internal_message = cast(TransactionMessage, internal_message)
-                self._handle_tx_message(internal_message)
-            else:
-                logger.warning(
-                    "Cannot handle a {} message.".format(type(internal_message))
-                )
-
-        # get new behaviours from the agent skills
-        for skill in self.resources.get_all_skills():
-            while not skill.skill_context.new_behaviours.empty():
-                new_behaviour = skill.skill_context.new_behaviours.get()
-                try:
-                    self.resources.behaviour_registry.register(
-                        (skill.skill_context.skill_id, new_behaviour.name),
-                        new_behaviour,
-                    )
-                except ValueError as e:
-                    logger.warning(
-                        "Error when trying to add a new behaviour: {}".format(str(e))
-                    )
-
-    def _handle_tx_message(self, tx_message: TransactionMessage):
-        """Handle transaction message from the Decision Maker."""
-        skill_callback_ids = tx_message.skill_callback_ids
-        for skill_id in skill_callback_ids:
-            handler = self.resources.handler_registry.fetch_internal_handler(skill_id)
-            if handler is not None:
-                logger.debug(
-                    "Calling handler {} of skill {}".format(type(handler), skill_id)
-                )
-                handler.handle(cast(Message, tx_message))
-            else:
-                logger.warning(
-                    "No internal handler fetched for skill_id={}".format(skill_id)
-                )
