+++ tests/test_aea.py
@@ -40,7 +40,7 @@ from aea.mail.base import Envelope
 from aea.protocols.base import Protocol
 from aea.protocols.default.message import DefaultMessage
 from aea.protocols.default.serialization import DefaultSerializer
-from aea.registries.base import Resources
+from aea.registries.resources import Resources
 from aea.skills.base import Skill
 
 from packages.fetchai.connections.local.connection import LocalNode, OEFLocalConnection
