+++ tests/test_docs/test_build_aea_programmatically/programmatic_aea.py
@@ -32,7 +32,7 @@ from aea.crypto.ledger_apis import LedgerApis
 from aea.crypto.wallet import Wallet
 from aea.identity.base import Identity
 from aea.protocols.base import Protocol
-from aea.registries.base import Resources
+from aea.registries.resources import Resources
 from aea.skills.base import Skill
 
 ROOT_DIR = "./"
