+++ tests/test_docs/test_cli_vs_programmatic_aeas/programmatic_aea.py
@@ -33,7 +33,7 @@ from aea.crypto.ledger_apis import LedgerApis
 from aea.crypto.wallet import Wallet
 from aea.identity.base import Identity
 from aea.protocols.base import Protocol
-from aea.registries.base import Resources
+from aea.registries.resources import Resources
 from aea.skills.base import Skill
 
 from packages.fetchai.connections.oef.connection import OEFConnection
